Switch demo to working on a big matrix
Vary (via a variable parameter) the matrix size. For now only do scalar
transpose, since it is the only thing that works.

Add scalar multiplication, but there are issues with it: a bug is
blocking nested loop and the pointer parameters at this point a mushed
together (as a potential workaround that did not work out).